requirements.txt
aiofiles==24.1.0
aiogram==3.15.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.11
aioredis==2.0.1
aiosignal==1.3.1
aiosqlite==0.20.0
alembic==1.14.0
amqp==5.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
APScheduler==3.11.0
async-timeout==5.0.1
asyncpg==0.30.0
attrs==24.2.0
babel==2.16.0
bcrypt==4.2.1
betterlogging==1.0.0
billiard==4.2.1
cachetools==5.5.0
celery==5.5.3
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
dnspython==2.7.0
email_validator==2.2.0
environs==11.0.0
et_xmlfile==2.0.0
fastapi==0.115.4
fastapi-cli==0.0.5
frozenlist==1.5.0
google-api-core==2.23.0
google-api-python-client==2.154.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.66.0
greenlet==3.1.1
gspread==6.1.4
h11==0.14.0
httpcore==1.0.6
httplib2==0.22.0
httptools==0.6.4
httpx==0.27.2
idna==3.10
ImageHash==4.3.2
iso8601==2.1.0
kombu==5.5.4
magic-filter==1.0.12
Mako==1.3.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.23.1
mdurl==0.1.2
multidict==6.1.0
numpy==2.1.3
oauthlib==3.2.2
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pendulum==3.0.0
pillow==11.3.0
prompt_toolkit==3.0.52
propcache==0.2.0
proto-plus==1.25.0
protobuf==5.29.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
pyparsing==3.2.0
pypika-tortoise==0.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
pytz==2024.2
PyWavelets==1.9.0
PyYAML==6.0.2
redis==6.4.0
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.9.4
rsa==4.9
scipy==1.16.2
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.0
text-unidecode==1.3
time-machine==2.16.0
tortoise-orm==0.22.2
typer==0.13.0
typing_extensions==4.12.2
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.1.1
urllib3==2.5.0
uvicorn==0.32.0
# uvloop==0.21.0
vine==5.1.0
watchfiles==0.24.0
wcwidth==0.2.13
websockets==14.0
yarl==1.18.0
